+++ src/scripts/application.js
@@ -1,11 +1,11 @@
 import { Application } from 'stimulus-repo/packages/stimulus'
-import { MathController } from './math_controller.js'
-import { PracticeController } from './practice_controller'
-import { QuizController } from './quiz_controller'
-import { EquationController } from './equation_controller'
+import { MathController } from './controllers/math_controller.js'
+import { PracticeController } from './controllers/practice_controller'
+import { QuizController } from './controllers/quiz_controller'
+import { EquationController } from './controllers/equation_controller'
 import { library, dom } from '@fortawesome/fontawesome-svg-core'
 import { faAppleAlt, faPencilRuler } from '@fortawesome/free-solid-svg-icons'
-import { NavigationBarController } from './navigation_bar_controller'
+import { NavigationBarController } from './controllers/navigation_bar_controller'
 
 require('../styles.scss')
