+++ src/scripts/controllers/equation_controller.js
