+++ src/scripts/controllers/navigation_bar_controller.js
